Cambio vida y danio a Double

+++ src/main/scala/clases/Monstruo.scala
@@ -5,10 +5,10 @@ import mixins.luchador.Luchador
 
 class Monstruo(vidaInicial :Int, var horror: Int = 1) extends Personaje with Luchador {
 
-  override var vida: Int = vidaInicial
+  override var vida: Double = vidaInicial
 
   //Propósito: cuando un mostruo ataca siempre ocaciona 1 de daño.
-  override def danio(): Int = 1
+  override def danio(): Double = 1
 
   def ocacionarHorror(investigador: Investigador): Unit = {
     investigador.recibirHorror(this.horror)
